Replace addressBook variables to eduTrack
