+++ src/test/java/seedu/address/model/UserPrefsTest.java
@@ -13,7 +13,7 @@ public class UserPrefsTest {
     }
 
     @Test
-    public void setAddressBookFilePath_nullPath_throwsNullPointerException() {
+    public void setEduTrackFilePath_nullPath_throwsNullPointerException() {
         UserPrefs userPrefs = new UserPrefs();
         assertThrows(NullPointerException.class, () -> userPrefs.setEduTrackFilePath(null));
     }
